install: minor improvements
1. It no longers does exec $SHELL after changing shell. That does not
work since the script would halt from that point. The problem it was
trying to solve is to reload the environment after setting up zsh. This
is useful for the vim Go plugin to install things in the right GOBIN
dir.

2. It does not try to kill apps that are not running. This would also
halt the script since killall would have a non-zero exit code.

3. Move the Solarized Dark Terminal theme into the config dir. This is
good because it is a config really. And also because at some point the
script may (today it is YouCompleteMe setup) change directories and we
should not rely on relative paths.

4. The file used a mix of tabs and spaces. Now it is consistent.